addidng edge cases for update present location

+++ sendapi/routes/parcel.py
@@ -89,11 +89,15 @@ def fetch_all_parcels():
 
 
 @app.route("/api/v1/parcels/<int:parcelId>", methods=['GET'])
+@jwt_required
 def fetch_specific_parcel(parcelId):
     """ 
         Function fetch specific parcel
         :return parcel:
     """
+    userId=get_jwt_identity()
+    if not userId:
+        return({"message":"Missing authentication header"}),401
     parcel = Parcel()
     single_parcel = parcel.get_single_parcel(parcelId)
     if not single_parcel:
@@ -110,11 +114,15 @@ def cancel_specific_parcel(parcelId):
     """
     userid = get_jwt_identity()
     parcel = Parcel()
-    if not parcel.get_single_parcel(parcelId):
+    Parcel_to_edit=parcel.get_single_parcel(parcelId)
+    if not Parcel_to_edit:
         return jsonify({"message": " parcel doesnot exist"}), 404
-    parcel_owner_id = parcel.get_single_parcel(parcelId)['userid']
+    parcel_owner_id = Parcel_to_edit['userid']
     if parcel_owner_id != userid:
-        return jsonify({"message": "You can only cancel an order you created"}), 400
+       return jsonify({"message": "You can only cancel an order you created"}), 400
+    parcel_status =Parcel_to_edit['status']
+    if parcel_status=='Cancelled' or parcel_status=='Delivered':
+        return jsonify({"message":"Cant update a cancelled or deivered order "}),400
     parcel.cancel_parcel(parcelId)
     return jsonify({"message": "Your parcel order has been cancelled"}), 200
 
@@ -126,19 +134,29 @@ def update_present_location(parcelId):
         Function to update present location
         :return success message:
     """
+    request_data = request.get_json(force=True)
+    newlocation = request_data['new location']
+    if not get_jwt_identity():
+        return jsonify({"message": "Some fields are missing"}), 400
+    if len(request_data.keys()) != 1:
+        return jsonify({"message": "Some fields are missing"}), 400
+    validate_input = Validator().validate_string_input(newlocation)
+    if not validate_input:
+        return jsonify({"message": "The new location should be a none empty string"}),400
     userid = get_jwt_identity()
     parcel = Parcel()
     user = User()
     editor = user.get_user_by_id(userid)['role']
     if editor != 'admin':
         return jsonify({"message": "You can only update the present location if you are an admin"}), 400
-    request_data = request.get_json(force=True)
-    if len(request_data.keys()) != 1:
-        return jsonify({"message": "Some fields are missing"}), 400
-    newlocation = request_data['new location']
-    validate_input = Validator().validate_string_input(newlocation)
-    if not validate_input:
-        return jsonify({"message": "The new location should be a none empty string"})
+    Parcel_to_edit=Parcel().get_single_parcel(parcelId)
+    if not Parcel_to_edit or Parcel_to_edit['status']=='Cancelled':
+        return ({"message":"The parcel you are tring to edit doesnt exist"}),400
+    if Parcel_to_edit['present_location']==newlocation:
+        return jsonify({"message":"Present location is upto date"}),400
+    if newlocation==Parcel_to_edit['destination']:
+        Parcel_to_edit['status']='Delivered'
+        return jsonify({"message":"order delivered"}),200
     parcel.update_present_location(parcelId, newlocation)
     parcel = parcel.get_single_parcel(parcelId)
     return jsonify({"message": "Your location has been updated ", "updated parcel": parcel}), 200
@@ -154,9 +172,13 @@ def update_destination(parcelId):
     userid = get_jwt_identity()
     parcel = Parcel()
     user = User()
+    if not userid:
+        return jsonify({"message":"Authorised access"}),401
+
     editor = user.get_user_by_id(userid)['role']
     parcel = parcel.get_single_parcel(parcelId)
-    if not parcel:
+
+    if not parcel or parcel['status']=='Cancelled':
         return jsonify({"message": "The parcel you are editing doesnt exist"}), 404
     if not editor:
         return jsonify({"message": "You are not a registered user of the system"}), 401
@@ -170,7 +192,6 @@ def update_destination(parcelId):
     if not validate_destination:
         jsonify({"message": "destination must be a non empty string"}), 400
     Parcel().update_destination(parcelId, destination)
-    # parcel.update_destination(parcelId, destination)
     parcel = Parcel().get_single_parcel(parcelId)
     return jsonify({"message": "Your destination has been updated ", "updated parcel": parcel}), 200
 
@@ -183,8 +204,13 @@ def delete_parcel(parcelId):
         :return success message:
     """
     userid = get_jwt_identity()
+    if not userid:
+        return jsonify({"message":"Authorised access"}),401
     user = User()
     parcel = Parcel()
+    parcel_to_delete = parcel.get_single_parcel(parcelId)
+    if not parcel_to_delete or parcel_to_delete['status']=='Deleted':
+        return jsonify({"message":"Order doesnt exist"}),400
     editor = user.get_user_by_id(userid)['role']
     if editor == 'admin':
         parcel.delete_parcel(parcelId)
